amalgamated playertracker and entityplayerhandler to reduce the number of event listeners
